[Build] remove dependency jars and .classpath and lock snapshots (#193)
Dependencies are handled by Maven automatically, there is no need to add
and manage the jars manually. Furthermore they are not even considered
by Maven. This can lead to inconsistencies between the build and the
IDE.

This allows also to remove the Eclipse .classpath file from the
symja_android_library module since that file is generated automatically
when the Maven-project is imported and it is required.

Also lock the snapshot dependencies in order to avoid compilation errors
if they change their API in a new snapshot version, so it is still
possible to use new snapshot versions in a controlled manner.