Clean up Maven structure part 1 (#237)
* [Build] Format, re-order and clean-up POMs (#193)

- Format all pom.xml files
- update xsi:schemaLocation
- Change the order of elements to comply with the order recommended by
Maven:
https://maven.apache.org/developers/conventions/code.html#pom-code-convention

To sort the pom elements, use the following command (with subsequent
Eclipse XML-editor formatting:

"mvn com.github.ekryd.sortpom:sortpom-maven-plugin:sort
-Dsort.createBackupFile=false -Dsort.nrOfIndentSpace=-1
-Dsort.indentSchemaLocation=true -Dsort.sortPlugins=groupId,artifactId
-Dsort.sortDependencies=scope,groupId,artifactId"

* [Build] improve plug-in management and update plug-in versions (#193)

- Remove resource filtering where no resources are filtered at all or
where the resources did not contain any replaceable variables. Which in
the end were all occurrences.

- Make explicit call of maven-compiler-plugin in the root pom.xml, which
is implicitly called anyways and remove it from child poms.

* [Build] generated java-doc and source for all modules regularly (#193)

* [Build] improve dependency management (#193)

- versions of each dependency is now managed in the dependencyManagement
section of the root pom.xml
	- this ensures consistent versions throughout the library

- Add JUnit 5 test-dependency (plus Vintage for Junit4) for all modules
in test scope

* [Build] remove dependency jars and .classpath and lock snapshots (#193)

Dependencies are handled by Maven automatically, there is no need to add
and manage the jars manually. Furthermore they are not even considered
by Maven. This can lead to inconsistencies between the build and the
IDE.

This allows also to remove the Eclipse .classpath file from the
symja_android_library module since that file is generated automatically
when the Maven-project is imported and it is required.

Also lock the snapshot dependencies in order to avoid compilation errors
if they change their API in a new snapshot version, so it is still
possible to use new snapshot versions in a controlled manner.