[Tests] remove package test suites (preparation for #239)
The removed test-suites contained all tests of their package (and one
even of other packages).
But those test-suites are not in used within the build and modern day
IDE can execute all test-clases of a package as simple as a single test
class. So there is no need for those test suites.

And removing the test suites also simplifies the transition to JUnit-5.